Test for ImageStack Volumes
Implemented a similar test for an image stack volume as was done for
regular volumes.

ImageStack Volume:
Removed translations - these will be added back in later
Changed np.matmul to np.multiply since the latter is element wise and
the former returns the scalar that you'd expect from regular matrix
multiplication.
Added mask_data as None - should be changed in the future
Fixed col_range and row_range in populator to be compatable with python3
 by using lists around the maps.